+++ CHANGELOG.md
@@ -3,6 +3,15 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased]
 
+## [1.1] -- 2021-01-23
+### Added
+- Non associative operators
+- Error handling on partially applied operators (which raises a
+  `Stream.Failure`)
+### Changed
+- One function `get` for operators in API
+- `make_appl` does not use the table of operators
+
 ## [1.0.1] -- 2021-01-16
 ### Fixed
 - Correct OCaml dependency